PromptFlow/legal-main-flow/docker/flow/requirements.txt
promptflow
openai
typing
azure-search-documents
promptflow-tools